% Generated by roxygen2 (4.0.1): do not edit by hand
\name{BO2dFix}
\alias{BO2dFix}
\title{Fix 2d-ordinations}
\usage{
BO2dFix(BO, eta, etaMean)
}
\arguments{
\item{BO}{2d Bayesian ordination}

\item{eta}{optional results of \code{BOeta(BO)}}

\item{etaMean}{optional posterior mean of \code{eta}}
}
\description{
Fix 2d-ordinations
}

