% Generated by roxygen2 (4.0.1): do not edit by hand
\name{BOeta}
\alias{BOeta}
\title{Compute the posterior of the linear predictor in a Bayesian
ordination}
\usage{
BOeta(BO)
}
\arguments{
\item{BO}{Results of \code{\link{getBO}}}
}
\value{
An array containing the posterior of the linear predictor
with dimensions relating to (1) MCMC iterations, (2) sites, and
(3) species
}
\description{
Compute the posterior of the linear predictor in a Bayesian
ordination
}

