% Generated by roxygen2 (4.0.1): do not edit by hand
\name{BOflip}
\alias{BOflip}
\title{Flip axes in a Bayesian ordination}
\usage{
BOflip(BO, refSites)
}
\arguments{
\item{BO}{Results of \code{\link{getBO}}}

\item{refSites}{Vector of reference sites for each axis}
}
\value{
A Bayesian ordination with flipped axes
}
\description{
Flip axes in a Bayesian ordination
}

