% Generated by roxygen2 (4.0.1): do not edit by hand
\name{BOswitch}
\alias{BOswitch}
\title{Switch axes labels}
\usage{
BOswitch(BO)
}
\arguments{
\item{BO}{Results of \code{\link{getBO}}}
}
\value{
A Bayesian ordination with switched axes
}
\description{
Attempts to switch axis labels such that more highly correlated
MCMC samples are labeled as the same axis.  Only for two-axis
models.
}

