% Generated by roxygen2 (4.0.1): do not edit by hand
\name{dimIRT}
\alias{axesIRT}
\alias{dimIRT}
\alias{itersIRT}
\alias{sitesIRT}
\alias{speciesIRT}
\title{The dimensions of an item response theory mcmc}
\usage{
dimIRT(mcmc, index)

itersIRT(mcmc, index)

sitesIRT(mcmc, index)

speciesIRT(mcmc, index)

axesIRT(mcmc, index)
}
\arguments{
\item{mcmc}{An \code{\link{mcmc}} object from a fitted IRT model.}

\item{index}{Optional (not really needed)}
}
\description{
The dimensions of an item response theory mcmc
}

