% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{ecoBayesIRT}
\alias{ecoBayesIRT}
\alias{ecoBayesIRT-package}
\title{Helper functions for ecological Bayesian item response theory in R
using MCMCpack}
\description{
Steps: (1) Fit a model using functions from the \code{MCMCpack}
package such as \code{\link{MCMCirt1d}} or
\code{\link{MCMCirtKd}}.  (2) Do diagnostics and convergence
checks on the resulting MCMC objects.  (3) Separate the three
types of parameters (site scores, \code{x}, species intercepts,
\code{a}, and species scores, \code{b}), by passing the MCMC
through the \code{\link{getBO}} function.  (4) Use the other
functions of this package to analyze the results
(\code{\link{BOeta}}; \code{\link{BOflip}}; \code{\link{BOp}};
\code{\link{BOrotate}}; \code{\link{BOswitch}}).
}
\examples{
## simulate data
nSite <- 25
nSpec <- 10
nIter <- 10000
set.seed(1)
x <- rnorm(nSite)
a <- rnorm(nSpec)
b <- rnorm(nSpec)
eta <- cbind(1, x) \%*\% rbind(a, b)
e <- matrix(rnorm(nSite*nSpec), nSite, nSpec)
p <- pnorm(eta + e)
Y <- matrix(rbinom(nSite*nSpec, 1, p), nSite, nSpec)
rownames(Y) <- letters[1:nSite]
colnames(Y) <- LETTERS[1:nSpec]

## fit model
Yirt <- MCMCirt1d(Y, mcmc = nIter, store.item = TRUE)

## getBO
BO <- getBO(Yirt, Y)

## fix identifiability
(refSite <- findRefSite(BO))
BO <- BOflip(BO, refSite)

## get fitted values
pFit <- BOp(BO)
pFitMean <- apply(pFit, c(2, 3), mean)
boxplot(pFitMean ~ Y)

## ordination plot
ordInt <- t(apply(BO$x, c(2, 3), quantile, probs = c(0.025, 0.975))[,,1])
}

