% Generated by roxygen2 (4.0.1): do not edit by hand
\name{extractIRT}
\alias{extractIRT}
\title{only meant to be used as interactive functions
(see http://adv-r.had.co.nz/Computing-on-the-language.html#calling-from-another-function)}
\usage{
extractIRT(mcmc, condition, index)
}
\description{
only meant to be used as interactive functions
(see http://adv-r.had.co.nz/Computing-on-the-language.html#calling-from-another-function)
}

