% Generated by roxygen2 (4.0.1): do not edit by hand
\name{findModes}
\alias{findModes}
\title{Differentiate between specific forms of uni- and bi-modality}
\usage{
findModes(den, tol = 0.01)
}
\arguments{
\item{den}{output of \code{\link{logspline}}}

\item{tol}{tolerance when comparing modes with zero}
}
\value{
A vector with the location of one or two modes
}
\description{
If there is one mode on either side of zero, then this function
attempts to return the location of those two modes.  If there is
only one mode, the function attempts to return the location of
that mode.  The algorithm is based on optimizing over the negative
numbers, then over the positive numbers, and seeing if either is
on the boundary.  This is not well tested.
}
\examples{
set.seed(1)
x <- rnorm(1000, c(-2, 2), 1)
den <- logspline(x)
plot(den)
abline(v = findModes(den))
}

