% Generated by roxygen2 (4.0.1): do not edit by hand
\name{fitProbBins}
\alias{fitProbBins}
\title{Compare binary data to binned probabilistic expectations}
\usage{
fitProbBins(y, p, length = 11)
}
\arguments{
\item{y}{vector of binary data}

\item{p}{vector of probabilities}

\item{length}{number of bins}
}
\value{
\code{data.frame} with observed proportions and expected
proportions in evenly spaced bins.
}
\description{
Compare binary data to binned probabilistic expectations
}

