% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getBO}
\alias{getBO}
\title{Get parameters from a Bayesian ordination (getBO)}
\usage{
getBO(mcmc, Y, index, name = c("x", "a", "b"))
}
\arguments{
\item{mcmc}{An mcmc object}

\item{Y}{Data matrix that matches the mcmc object}

\item{index}{The results of \code{\link{BOindex}}}

\item{name}{What to extract
(site scores, x, intercepts, a, or species scores, b)}
}
\value{
extracted object
}
\description{
Get parameters from a Bayesian ordination (getBO)
}

