% Generated by roxygen2 (4.0.1): do not edit by hand
\name{give}
\alias{give}
\title{Give values to certain dimension of an R array}
\usage{
give(x, along, indices, value)
}
\arguments{
\item{x}{matrix or array to subset}

\item{along}{dimension to subset along}

\item{indices}{the indices to select}

\item{value}{the values to give}
}
\description{
Modelled after the \code{\link{take}} function in the \code{plyr}
package.
}

