% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mcmcFlip}
\alias{mcmcFlip}
\title{Flip axes in an mcmc object for a Bayesian ordination}
\usage{
mcmcFlip(mcmc, refSites)
}
\arguments{
\item{mcmc}{An \code{\link{mcmc}} object}

\item{refSites}{Vector of reference sites for each axis}
}
\value{
An \code{\link{mcmc}} object for a Bayesian ordination
with flipped axes
}
\description{
Flip axes in an mcmc object for a Bayesian ordination
}
\seealso{
\code{\link{BOflip}}
}

