% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot2dCors}
\alias{plot2dCors}
\title{Plot correlations between 2d indirect and 1d direct gradients}
\usage{
plot2dCors(xIndirect, xDirect, .seed = 1, plotArgs = list(las = 1, xlim =
  c(-1, 1), ylim = c(-1, 1), xlab = "", ylab = ""), nullPtArgs = list(col =
  rgb(1, 0, 0, 0.1)), xPtArgs = list(col = rgb(0, 0, 1, 0.1)),
  nullContArgs = list(col = "red", lwd = 3, cont = 95), xContArgs = list(col
  = "blue", lwd = 3, cont = 95))
}
\arguments{
\item{xIndirect}{posterior distribution of indirect gradient
values (MCMC iterations-by-sites-by-2 axes array)}

\item{xDirect}{direct gradient (vector with same length as second
dimension of \code{xIndirect})}

\item{.seed}{random seed for generation of the null correlations}

\item{plotArgs}{list of arguments for setting up the plotting region}

\item{nullPtArgs}{list of arguments for plotting correlations from
a null distribution}

\item{xPtArgs}{list of arguments for plotting correlations with direct gradient}

\item{nullContArgs}{list of arguments for plotting null distribution contour}

\item{xContArgs}{list of arguments for plotting contour}
}
\description{
Plot correlations between 2d indirect and 1d direct gradients
}

