% Generated by roxygen2 (4.0.1): do not edit by hand
\name{quantile2d}
\alias{quantile2d}
\title{Two-dimensional quantile plot}
\usage{
quantile2d(x, y, p, ...)
}
\arguments{
\item{x}{numeric vector}

\item{y}{numeric vector the same length as \code{y}}

\item{p}{probability}

\item{...}{other arguments to \code{\link{polygon}}}
}
\value{
plots a convex hull that contains a fraction \code{p} of
the \code{x}-\code{y} points.
}
\description{
Two-dimensional quantile plot
}

