% Generated by roxygen2 (4.0.1): do not edit by hand
\name{redundantPara}
\alias{dropRedundantPara}
\alias{redundantPara}
\title{Drop redundant parameters}
\usage{
redundantPara(mcmc)

dropRedundantPara(mcmc)
}
\arguments{
\item{mcmc}{An \code{\link{mcmc}} object from a fitted IRT model.}
}
\description{
Drop redundant parameters
}

