% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rgrad}
\alias{rgrad}
\title{Simulate from posterior distribution of indirect gradients}
\usage{
rgrad(n, mcmc)
}
\arguments{
\item{n}{number of replicates.}

\item{mcmc}{an \code{mcmc} object from an IRT.}
}
\description{
Simulate from posterior distribution of indirect gradients
}

