% Generated by roxygen2 (4.0.1): do not edit by hand
\name{swapAxesIRT}
\alias{flipAxesIRT}
\alias{swapAxesIRT}
\title{Swap and flip axes of an item response theory mcmc}
\usage{
swapAxesIRT(mcmc, ord, param = c("beta", "theta"))

flipAxesIRT(mcmc, mult, param = c("theta", "beta"))
}
\arguments{
\item{ord}{\code{itersIRT(mcmc)} by \code{axesIRT(mcmc)} matrix
with the orderings of the axes.}

\item{mult}{\code{itersIRT(mcmc)} by \code{axesIRT(mcmc)} matrix
\code{-1}/\code{1} multipliers for flipping the axes.}

\item{mcmc}{An \code{\link{mcmc}} object from a fitted IRT model.}

\item{param}{Which parameter to obtain/replace (\code{theta},
\code{beta}, or \code{alpha}).}
}
\description{
Swap and flip axes of an item response theory mcmc
}

