% Generated by roxygen2 (4.0.1): do not edit by hand
\name{apply2arrays}
\alias{apply2arrays}
\title{Returns a vector or array or list of values obtained by applying a
function to margins of two arrays or matrices.}
\usage{
apply2arrays(X, Y, XMARGIN, YMARGIN, FUN, ...)
}
\arguments{
\item{X,Y}{arrays (each potentially a matrix)}

\item{XMARGIN,YMARGIN}{vectors giving the subscripts which the
function will be applied over.  see \code{\link{apply}} for more
information. for the special case here, note that
\code{all(dim(X)[XMARGIN] == dim(Y)[YMARGIN])} must be
\code{TRUE}.}

\item{FUN}{the function to be applied.}

\item{...}{more arguments to \code{FUN}.}
}
\value{
a new array, or list if \code{\link{simplify2array}}
cannot do it's job.
}
\description{
Returns a vector or array or list of values obtained by applying a
function to margins of two arrays or matrices.
}
\examples{
dims <- c(4, 2, 3, 1, 2)
set.seed(1)
(A <- array(rnorm(prod(dims[-2])), dims[-2]))
(B <- array(rnorm(prod(dims[-(3:4)])), dims[-(3:4)]))
apply2arrays(A, B, c(1, 4), c(1, 3), `\%*\%`)
}

