% Generated by roxygen2 (4.0.1): do not edit by hand
\name{BOindex}
\alias{BOindex}
\alias{setIndex}
\alias{setIndex.mcmc}
\alias{setIndex.mcmc.list}
\title{Construct an index for a Bayesian ordination (BO) mcmc object}
\usage{
BOindex(mcmc)

setIndex(mcmc)

\method{setIndex}{mcmc}(mcmc)

\method{setIndex}{mcmc.list}(mcmc)
}
\arguments{
\item{mcmc}{An \code{mcmc} or \code{mcmc.list} object}
}
\value{
A data frame with one row per model parameter.  Each
row gives the (1) parameter type, (2) object associated with
that parameter (e.g. species name), and (3) the ordination
axis associated with that parameter (if it is an axis-specific
parameter).
}
\description{
Construct an index for a Bayesian ordination (BO) mcmc object
}

