% Generated by roxygen2 (4.0.1): do not edit by hand
\name{extractIRT}
\alias{a}
\alias{a<-}
\alias{b}
\alias{b<-}
\alias{extractIRT}
\alias{getAxisParamArray}
\alias{replaceIRT}
\alias{setAxisParamArray}
\alias{swapAxisParamArray}
\alias{x}
\alias{x<-}
\title{Extract and replace pieces of an item response theory mcmc}
\usage{
extractIRT(mcmc, condition, index)

replaceIRT(mcmc, condition, value, index)

a(mcmc, index)

b(mcmc, index)

x(mcmc, index)

a(mcmc) <- value

b(mcmc) <- value

x(mcmc) <- value

getAxisParamArray(mcmc, index, param = c("beta", "theta"))

setAxisParamArray(mcmc, index, param = c("beta", "theta"), value)

swapAxisParamArray(mcmc, index, param = c("beta", "theta"), ord)
}
\arguments{
\item{mcmc}{An \code{\link{mcmc}} object from a fitted IRT model.}

\item{condition}{An expression evaluating to a
\code{\link{logical}} vector the same length as the number of
columns (parameters) in \code{mcmc}.}

\item{index}{Optional (not really needed)}

\item{value}{Replacement value.}

\item{param}{Which parameter to obtain/replace (\code{theta},
\code{beta}, or \code{alpha}).}

\item{ord}{\code{itersIRT(mcmc)} by \code{axesIRT(mcmc)} matrix
with the orderings of the axes.}
}
\value{
Another \code{\link{mcmc}} object with (1) only the
extracted piece (for \code{extractIRT}) or (2) the replaced values
(for \code{replaceIRT}).
}
\description{
\code{extractIRT} and \code{replaceIRT} are only to be used in
interactive mode (i.e. not used within functions).  see
http://adv-r.had.co.nz/Computing-on-the-language.html#calling-from-another-function.
It is much safer and easier to read to use the \code{x}, \code{b},
and \code{a} functions, which are used to extract and replace all
gradient values, species slopes, and species intercepts (note
these functions return as arrays).  \code{getAxisParamArray} and
\code{setAxisParamArray} are mostly for internal use, because
usually \code{x}, \code{b}, and \code{a} will provide everything
someone might need.
}

