% Generated by roxygen2 (4.0.1): do not edit by hand
\name{pairComp}
\alias{pairComp}
\title{Pairwise comparisons among columns in binary or probability
matrices}
\usage{
pairComp(y)
}
\arguments{
\item{y}{binary or probability matrix}
}
\value{
4-d array with comparisons
}
\description{
Pairwise comparisons among columns in binary or probability
matrices
}

