% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rgrad}
\alias{rgrad}
\title{Simulate from posterior distribution of indirect gradients}
\usage{
rgrad(mcmc, plot = FALSE, ...)
}
\arguments{
\item{mcmc}{an \code{mcmc} object from an IRT.}

\item{plot}{should a biplot be plotted?}

\item{...}{additional arguments to \code{\link{biplot}}.}
}
\description{
Simulate from posterior distribution of indirect gradients
}

